Create ReadAndWriteFromScore
A dual-purpose file that reads the selection from the score, exports the output as a file, then imports a file, and imports the notes directly onto the page immediately after the selection. From the standpoint of the user, it will merely look as though the user had copied a selection and pasted it immediately after. The use of this is that the output file is read into the model, and the input file is the output from the model. I have yet to integrate the sorted (and working) write code into this main code block, but my read code works properly even without it being sorted.